\name{actuator.calibrate}
\title{
Actuator calibration experiment.
}
\section{SUMMARY}{
The \code{actuator.calibrate} data frame has 30 rows and 3 columns.
The actuator is a small propulsion unit used to maneuver space craft
in space. The actuator is fired using compressed air and can be
calibrated by the amount of pressure used.
}
\section{"DATA}{
This data frame contains the following columns:
}
\arguments{
\item{force}{
Force produced in pounds.
}
\item{press}{
Amount of pressure in psi.
}
\item{order}{
Actual order data was generated.
}}
\section{SOURCE}{
John Meyer
}
\seealso{
actuator
}
\keyword{FUNFITS}
% Converted by Sd2Rd version 0.2-a3.
