\name{actuator}
\title{
Actuator force experiment.
}
\section{SUMMARY}{
The \code{actuator} data frame has 16 rows and 6 columns.
The actuator is a small propulsion unit used to maneuver space craft
in space. The actuator is fired using compressed air and the factors
used in the experiment are the actuator used, the amount of pressure
used, the length of air supply line and the nozzle type.
}
\section{"DATA}{
This data frame contains the following columns:
}
\arguments{
\item{act}{
Actuator used in the experiment.
}
\item{press}{
Amount of pressure in psi.
}
\item{line}{
Length of air supply line in feet.
}
\item{nozzle}{
Nozzle type.
}
\item{force}{
Force produced in pounds.
}
\item{order}{
Actual order data was generated. 
}}
\section{SOURCE}{
John Meyer
}
\seealso{
actuator.calibrate,
}
\keyword{FUNFITS}
% Converted by Sd2Rd version 0.2-a3.
