\name{BD2}
\title{
}
\section{SUMMARY}{
The \code{BD2} data frame has 64 rows and 6 columns representing
a sequence of RSM designs with 3 design variables.
}
\section{"DATA}{
This data frame contains the following columns:
}
\arguments{
\item{expt}{
Experiment number.
}
\item{kpo4}{
phosphate concentration
}
\item{mgoac}{
magnesium acetate concentration
}
\item{dmso}{
dmso concentration
}
\item{amtb}{
response measurement adjusted by experiment controls 
}
\item{lamtb}{
log of adjusted response measurement
}}
\section{SOURCE}{
Becton Dickinson Research Center
Research Triangle Park, NC
}
\seealso{
nnreg, tps, krig
}
\keyword{FUNFITS}
% Converted by Sd2Rd version 0.2-a3.
