\name{climate}
\title{
Climate and geographical data for 50 largest US cities.


Information on 50 largest US cities.
List components:


lon: longitude


lat: latitude


jan: average minimum January temperature (degrees F)


rain: average rainfall in inches


city: city names


jul: average maximum July temperature


elev: Elevation above sea level in KW.






Reference: The Universal Almanac (1992), ed. John W. Wright, Andrews and
McNeel, Kansas City.
}
\keyword{slab}
% Converted by Sd2Rd version 0.2-a3.
