

\name{flame}
\title{
Response surface for a intensity 


Results from an analytical chemistry lab at NCSU. The intensity of
emitted light when a solution with Lithium is ionized in a flame. The
two independent variables are the fuel and O2 flow rates for the
flame. The goal is to identify flow rates that maximize the intensity.
The ridge in this surface may be due to stoichiometric ratio of how the
fuel and the oxygen combine to burn. A maxima is located near the edge
of the operating region but it was found that for setting of flow rates
outside the experimental region the mixture tended to explode!


Data was generously provided by Charles Boss and Kevin Buckley. 


}
\keyword{FUNFITS}
% Converted by Sd2Rd version 0.2-a3.
