

\name{minitri}
\title{
Mini triathlon results


Results from a mini triathlon sponsored by Bud Lite, held in
Cary, NC, June 1990. Times are in minutes for the male 30-34 group.
Components are:


swim:  (1/2 mile)


bike:  (15 miles)


run:   (4 miles)
}
\keyword{slab}
% Converted by Sd2Rd version 0.2-a3.
