\name{vgfast}
\title{
Fast variogram for replicated fields
}
\usage{
vgfast(data)
}
\arguments{
\item{data}{
A matrix where the columns are different locations  and the rows are 
independent realizations of the field. 
}}
\value{
A vector of average squared differences


}
\seealso{
vgram


\}
}
\keyword{spatial}
% Converted by Sd2Rd version 0.2-a3.
